pyproject.toml
[tool.poetry]
name = "django-project-default"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
uWSGI = "^2.0.19"
python-dotenv = "^0.15.0"
djangorestframework = "^3.12.2"
djangorestframework-jwt = "^1.11.0"
django-filter = "^2.4.0"
django-compressor = "^2.4"
Pillow = "^8.1.0"
channels-redis = "^3.2.0"
channels = "^3.0.3"
Markdown = "^3.3.3"
daphne = "^3.0.1"
django-cors-headers = "^3.7.0"
django-sass-processor = "^0.8.2"
libsass = "^0.20.1"
slackweb = "^1.0.5"
numpy = "^1.20.1"
mysqlclient = "^2.0.3"
stripe = "^2.55.2"
mecab-python3 = "^1.0.3"
gensim = "^3.8.3"
requests-oauthlib = "^1.3.0"
drf-yasg = "^1.20.0"
google-cloud-translate = "1.7.0"
drf-writable-nested = "^0.7.0"

[tool.poetry.dev-dependencies]
httpie = "^2.4.0"
flake8 = "^3.8.4"
yapf = "^0.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
